% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_school_year}
\alias{format_school_year}
\title{Format school year string}
\usage{
format_school_year(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Character string in YYYYYYYY format
}
\description{
Converts end year to school year string format (e.g., 2024 -> "20232024")
}
\keyword{internal}
