% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_years}
\alias{get_available_years}
\title{Get available years for Nebraska enrollment data}
\usage{
get_available_years()
}
\value{
Integer vector of available school year ends
}
\description{
Returns a vector of school year ends for which enrollment data is available.
Nebraska provides historical enrollment data from 2002-03 through present:
\itemize{
\item 2018-present: CSV files (MembershipByGradeRaceAndGender)
\item 2011-2017: TXT files with current column format
\item 2003-2010: TXT files with legacy column format (pre-2010 race categories)
}
}
\details{
Note: Years 2001-2002 (end_year 2002) and earlier have different data structures
and are not currently supported.
}
\examples{
get_available_years()
}
