% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_known_file_urls}
\alias{get_known_file_urls}
\title{Get known file URLs for historical years}
\usage{
get_known_file_urls()
}
\value{
Named list of URL vectors by end_year
}
\description{
Returns a list of known working URLs for each historical year.
These URLs are extracted from NDE's data archives page.
}
\keyword{internal}
