% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_directory.R
\name{get_raw_directory}
\alias{get_raw_directory}
\title{Download raw directory data from NDE}
\usage{
get_raw_directory(admin_type = "all")
}
\arguments{
\item{admin_type}{Type of administrator ("superintendents", "elementary_principals",
"middle_principals", "secondary_principals", "all")}
}
\value{
List with raw data frames for each administrator type requested
}
\description{
Downloads the Nebraska administrator contact list Excel file(s).
}
\keyword{internal}
