% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{map_grade_code}
\alias{map_grade_code}
\title{Map NDE grade codes to standard format}
\usage{
map_grade_code(grade_code)
}
\arguments{
\item{grade_code}{Vector of NDE grade codes}
}
\value{
Vector of standardized grade levels
}
\description{
Nebraska uses codes like PK, KN, 01-12
}
\keyword{internal}
