% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_district_code}
\alias{parse_district_code}
\title{Parse Nebraska district code}
\usage{
parse_district_code(co_dist)
}
\arguments{
\item{co_dist}{Character vector of CO_DIST codes}
}
\value{
List with county and district components
}
\description{
Nebraska uses CO_DIST format (county-district) like "01-0003"
}
\keyword{internal}
