% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_school_code}
\alias{parse_school_code}
\title{Parse Nebraska school code}
\usage{
parse_school_code(co_dist_sch)
}
\arguments{
\item{co_dist_sch}{Character vector of CO_DIST_SCH codes}
}
\value{
List with county, district, and school components
}
\description{
Nebraska uses CO_DIST_SCH format (county-district-school) like "01-0003-001"
}
\keyword{internal}
